extracted mechanic into separate file
